+++ src/document.js
@@ -305,6 +305,7 @@ const documentsMain = {
 								label: t('richdocuments', 'Open in local editor'),
 								hint: t('richdocuments', 'Open in local editor'),
 								insertBefore: 'print',
+								accessKey: '2',
 							})
 						}